requirements.txt
pandas
streamlit
numpy
XlsxWriter
deta
python-dotenv
streamlit-nested-layout
streamlit-option-menu